+++ src/components/Editor/Editor.jsx
@@ -13,6 +13,7 @@
  import React, { useContext, useState, useRef, useEffect } from "react";
  import MonacoEditor from "react-monaco-editor";
  import ediTDorContext from "../../context/ediTDorContext";
+ import { changeBetweenTd} from '../../util';
 
 const mapping = require("../../assets/mapping.json");
 
@@ -228,45 +229,7 @@ const JSONEditorComponent = (props) => {
 
    const changeLinkedTd = async () =>{
       let href = document.getElementById("linkedTd").value;
-      context.setFileHandle(undefined)
-      if (context.linkedTd[href]["kind"]==="file"){
-        try{
-          if(context.isModified){
-            var  writable = await context.fileHandle.createWritable();
-          }
-          } catch(e){
-              document.getElementById("linkedTd").value=href
-              let fileHandle=context.linkedTd[href]
-              const file = await fileHandle.getFile();
-              const td=JSON.parse(await file.text())
-              let offlineTd=JSON.stringify(td, null, 2)
-              context.setFileHandle(fileHandle)
-              context.updateOfflineTD(offlineTd)
-              context.updateIsModified(false)
-
-          }
-          if(writable){
-            await writable.write(context.offlineTD)
-            await writable.close()
-          }
-          document.getElementById("linkedTd").value=href
-          let fileHandle=context.linkedTd[href]
-          const file = await fileHandle.getFile();
-          const td=JSON.parse(await file.text())
-          let offlineTd=JSON.stringify(td, null, 2)
-          context.setFileHandle(fileHandle)
-          context.updateOfflineTD(offlineTd)
-          context.updateIsModified(false)
-      }
-      // If we create a TD using the New button then we don't have a file handler
-      // In that case the entry in linkedTd is not a file handler but a Thing Description Json 
-      else if(Object.keys(context.linkedTd[href]).length){
-        const td=context.linkedTd[href]
-        let offlineTd=JSON.stringify(td, null, 2)
-        context.updateOfflineTD(offlineTd)
-        context.updateIsModified(false)
-
-      }
+      changeBetweenTd(context,href)
     }
   return (
      <>
