+++ src/components/TDViewer/Link.jsx
@@ -14,6 +14,7 @@
  import "../../assets/main.css"
  import ediTDorContext from "../../context/ediTDorContext";
  import { Trash2,Info } from "react-feather";
+ import { changeBetweenTd} from '../../util';
 
  export default function Link(props) {
 
@@ -31,45 +32,7 @@
      }
      const infoLink = async (e) =>{
       let href=e.link.href
-      context.setFileHandle(undefined)
-      if (context.linkedTd[href]["kind"]==="file"){
-        try{
-          if(context.isModified){
-            var  writable = await context.fileHandle.createWritable();
-          }
-          } catch(e){
-              document.getElementById("linkedTd").value=href
-              let fileHandle=context.linkedTd[href]
-              const file = await fileHandle.getFile();
-              const td=JSON.parse(await file.text())
-              let offlineTd=JSON.stringify(td, null, 2)
-              context.setFileHandle(fileHandle)
-              context.updateOfflineTD(offlineTd)
-              context.updateIsModified(false)
-
-          }
-          if(writable){
-            await writable.write(context.offlineTD)
-            await writable.close()
-          }
-          document.getElementById("linkedTd").value=href
-          let fileHandle=context.linkedTd[href]
-          const file = await fileHandle.getFile();
-          const td=JSON.parse(await file.text())
-          let offlineTd=JSON.stringify(td, null, 2)
-          context.setFileHandle(fileHandle)
-          context.updateOfflineTD(offlineTd)
-          context.updateIsModified(false)
-        }
-        // If we create a TD using the New button then we don't have a file handler
-        // In that case the entry in linkedTd is not a file handler but a Thing Description Json 
-        else if(Object.keys(context.linkedTd[href]).length !== 0){
-          document.getElementById("linkedTd").value=href
-          const td=context.linkedTd[href]
-          let offlineTd=JSON.stringify(td, null, 2)
-          context.updateOfflineTD(offlineTd)
-          context.updateIsModified(false)
-        }
+      changeBetweenTd(context,href)
      }
      return (
         <div className="flex flex-row items-center justify-start h-10 w-full bg-formBlue rounded-md px-4 mt-2 bg-opacity-75 border-2 border-formBlue">
