+++ src/components/TDViewer/TDViewer.jsx
@@ -13,7 +13,7 @@
 import React, { useContext,useEffect } from 'react';
 import '../../assets/main.css';
 import ediTDorContext from '../../context/ediTDorContext';
-import { buildAttributeListObject, separateForms } from '../../util';
+import { buildAttributeListObject, separateForms,changeBetweenTd } from '../../util';
 import { AddFormDialog } from "../Dialogs/AddFormDialog";
 import { AddLinkTdDialog } from '../Dialogs/AddLinkTdDialog';
 import { InfoIconWrapper } from '../InfoIcon/InfoIcon';
@@ -169,49 +169,9 @@ export default function TDViewer() {
                                 rotate: true,
                                 action: async function(evt,elementView, buttonView) {
                                     let href=elementView.model.get("href")
-                                    context.setFileHandle(undefined)
-                                    if (context.linkedTd[href]["kind"]==="file"){
-                                        try{
-                                            if(context.isModified){
-                                                 var  writable = await context.fileHandle.createWritable();
-                                            }
-                                        } catch(e){
-                                            document.getElementById("linkedTd").value=href
-                                            let fileHandle=context.linkedTd[href]
-                                            const file = await fileHandle.getFile();
-                                            const td=JSON.parse(await file.text())
-                                            let offlineTd=JSON.stringify(td, null, 2)
-                                            context.setFileHandle(fileHandle)
-                                            context.updateOfflineTD(offlineTd)
-                                            context.updateIsModified(false)
-
-                                        }
-                                        if(writable){
-                                            await writable.write(context.offlineTD);
-                                            await writable.close();
-                                        }
-                                        document.getElementById("linkedTd").value=href
-                                        let fileHandle=context.linkedTd[href]
-                                        const file = await fileHandle.getFile();
-                                        const td=JSON.parse(await file.text())
-                                        let offlineTd=JSON.stringify(td, null, 2)
-                                        context.setFileHandle(fileHandle)
-                                        context.updateOfflineTD(offlineTd)
-                                        context.updateIsModified(false)
-
-                                      }
-                                      // If we create a TD using the New button then we don't have a file handler
-                                      // In that case the entry in linkedTd is not a file handler but a Thing Description .json 
-                                      else if(Object.keys(context.linkedTd[href]).length !== 0){
-                                        document.getElementById("linkedTd").value=href
-                                        const td=context.linkedTd[href]
-                                        let offlineTd=JSON.stringify(td, null, 2)
-                                        context.updateOfflineTD(offlineTd)
-                                        context.updateIsModified(false)
-
-                                      }
-                                    }
-                                });
+                                    changeBetweenTd(context,href)
+                                }
+                            });
                     let toolsView = new joint.dia.ToolsView({
                         tools: [removeButton,infoButton]
                     });
